+++ bindings/node/lib/implementations/tokenizers/base.tokenizer.ts
@@ -3,10 +3,11 @@ import { promisify } from "util";
 import { PostProcessor } from "../../bindings/post-processors";
 import {
   AddedToken,
+  EncodeInput,
+  EncodeOptions,
+  InputSequence,
   PaddingConfiguration,
   PaddingOptions,
-  TokenizedSequence,
-  TokenizedSequenceWithOffsets,
   Tokenizer,
   TruncationConfiguration,
   TruncationOptions
@@ -26,6 +27,18 @@ export class BaseTokenizer<TConfig extends object> {
      */
     readonly configuration: Readonly<TConfig>
   ) {}
+
+  /**
+   * Instantiate a new Tokenizer from the given file
+   * @param path Path to a file containing a Tokenizer
+   */
+  static fromFile = Tokenizer.fromFile;
+
+  /**
+   * Instantiate a new Tokenizer from the given JSON string
+   * @param s A JSON string representation of the Tokenizer
+   */
+  static fromString = Tokenizer.fromString;
 
   /**
    * Truncation configuration if enabled, `null` otherwise.
@@ -76,15 +89,15 @@ export class BaseTokenizer<TConfig extends object> {
    *
    * @param sequence The sequence to encode
    * @param [pair] The optional pair sequence
-   * @param [addSpecialTokens=true] Whether to add the special tokens while encoding
+   * @param [options] Some options to customize the encoding
    */
   async encode(
-    sequence: string,
-    pair?: string,
-    addSpecialTokens = true
+    sequence: InputSequence,
+    pair?: InputSequence,
+    options?: EncodeOptions
   ): Promise<Encoding> {
     const encode = promisify(this.tokenizer.encode.bind(this.tokenizer));
-    const rawEncoding = await encode(sequence, pair ?? null, addSpecialTokens);
+    const rawEncoding = await encode(sequence, pair ?? null, options ?? null);
     return new Encoding(rawEncoding);
   }
 
@@ -93,50 +106,14 @@ export class BaseTokenizer<TConfig extends object> {
    *
    * @param sequences A list of sequences or pair of sequences.
    * The list can contain both at the same time.
-   * @param [addSpecialTokens=true] Whether to add the special tokens while encoding
+   * @param [options] Sope options to customize the encoding
    */
   async encodeBatch(
-    sequences: (string | [string, string])[],
-    addSpecialTokens = true
+    sequences: EncodeInput[],
+    options?: EncodeOptions
   ): Promise<Encoding[]> {
     const encodeBatch = promisify(this.tokenizer.encodeBatch.bind(this.tokenizer));
-    const rawEncodings = await encodeBatch(sequences, addSpecialTokens);
-    return rawEncodings.map(e => new Encoding(e));
-  }
-
-  /**
-   * Encode the given tokens sequence
-   * @param sequence A sequence of tokens to encode.
-   * If the sequence is a `TokenizedSequence`, offsets will be automatically generated,
-   * making the hypothesis that all the tokens in the sequence are contiguous in the original string
-   * @param [typeId=0] The type id of the given sequence. Defaults to 0.
-   * @since 0.6.0
-   */
-  async encodeTokenized(
-    sequence: TokenizedSequence | TokenizedSequenceWithOffsets,
-    typeId?: number
-  ): Promise<Encoding> {
-    const encode = promisify(this.tokenizer.encodeTokenized.bind(this.tokenizer));
-    const rawEncoding = await encode(sequence, typeId);
-    return new Encoding(rawEncoding);
-  }
-
-  /**
-   * Encode the given tokens sequences
-   * @param sequences A list of sequences to encode.
-   * If a sequence is a `TokenizedSequence`, offsets will be automatically generated,
-   * making the hypothesis that all the tokens in the sequence are contiguous in the original string
-   * @param [typeId=0] The type id of the given sequences. Defaults to 0.
-   * @since 0.6.0
-   */
-  async encodeTokenizedBatch(
-    sequences: (TokenizedSequence | TokenizedSequenceWithOffsets)[],
-    typeId?: number
-  ): Promise<Encoding[]> {
-    const encodeBatch = promisify(
-      this.tokenizer.encodeTokenizedBatch.bind(this.tokenizer)
-    );
-    const rawEncodings = await encodeBatch(sequences, typeId);
+    const rawEncodings = await encodeBatch(sequences, options);
     return rawEncodings.map(e => new Encoding(e));
   }
 
@@ -262,6 +239,23 @@ export class BaseTokenizer<TConfig extends object> {
    */
   setPostProcessor(processor: PostProcessor): void {
     return this.tokenizer.setPostProcessor(processor);
+  }
+
+  /**
+   * Save the Tokenizer as JSON to the given path
+   * @param path Path to the JSON file to write
+   * @param [pretty=false] Whether the JSON string should be prettified
+   */
+  save(path: string, pretty?: boolean): void {
+    return this.tokenizer.save(path, pretty);
+  }
+
+  /**
+   * Get a serialized JSON version of the Tokenizer as a string
+   * @param [pretty=false] Whether the JSON string should be prettified
+   */
+  toString(pretty?: boolean): string {
+    return this.tokenizer.toString(pretty);
   }
 }