+++ bindings/python/src/decoders.rs
@@ -5,9 +5,10 @@ use super::utils::Container;
 use pyo3::exceptions;
 use pyo3::prelude::*;
 use pyo3::types::*;
+use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use tk::tokenizer::Result;
 
-#[pyclass(dict)]
+#[pyclass(dict, module = "tokenizers.decoders")]
 pub struct Decoder {
     pub decoder: Container<dyn tk::tokenizer::Decoder>,
 }
@@ -21,12 +22,41 @@ impl Decoder {
         })
     }
 
+    fn __getstate__(&self, py: Python) -> PyResult<PyObject> {
+        let data = self
+            .decoder
+            .execute(|decoder| serde_json::to_string(&decoder))
+            .map_err(|e| {
+                exceptions::Exception::py_err(format!(
+                    "Error while attempting to pickle Decoder: {}",
+                    e.to_string()
+                ))
+            })?;
+        Ok(PyBytes::new(py, data.as_bytes()).to_object(py))
+    }
+
+    fn __setstate__(&mut self, py: Python, state: PyObject) -> PyResult<()> {
+        match state.extract::<&PyBytes>(py) {
+            Ok(s) => {
+                self.decoder =
+                    Container::Owned(serde_json::from_slice(s.as_bytes()).map_err(|e| {
+                        exceptions::Exception::py_err(format!(
+                            "Error while attempting to unpickle Decoder: {}",
+                            e.to_string()
+                        ))
+                    })?);
+                Ok(())
+            }
+            Err(e) => Err(e),
+        }
+    }
+
     fn decode(&self, tokens: Vec<String>) -> PyResult<String> {
         ToPyResult(self.decoder.execute(|decoder| decoder.decode(tokens))).into()
     }
 }
 
-#[pyclass(extends=Decoder)]
+#[pyclass(extends=Decoder, module = "tokenizers.decoders")]
 pub struct ByteLevel {}
 #[pymethods]
 impl ByteLevel {
@@ -41,7 +71,7 @@ impl ByteLevel {
     }
 }
 
-#[pyclass(extends=Decoder)]
+#[pyclass(extends=Decoder, module = "tokenizers.decoders")]
 pub struct WordPiece {}
 #[pymethods]
 impl WordPiece {
@@ -71,7 +101,7 @@ impl WordPiece {
     }
 }
 
-#[pyclass(extends=Decoder)]
+#[pyclass(extends=Decoder, module = "tokenizers.decoders")]
 pub struct Metaspace {}
 #[pymethods]
 impl Metaspace {
@@ -109,7 +139,7 @@ impl Metaspace {
     }
 }
 
-#[pyclass(extends=Decoder)]
+#[pyclass(extends=Decoder, module = "tokenizers.decoders")]
 pub struct BPEDecoder {}
 #[pymethods]
 impl BPEDecoder {
@@ -147,6 +177,7 @@ impl PyDecoder {
     }
 }
 
+#[typetag::serde]
 impl tk::tokenizer::Decoder for PyDecoder {
     fn decode(&self, tokens: Vec<String>) -> Result<String> {
         let gil = Python::acquire_gil();
@@ -165,5 +196,25 @@ impl tk::tokenizer::Decoder for PyDecoder {
                 Err(Box::new(PyError::from("Error while calling `decode`")))
             }
         }
+    }
+}
+
+impl Serialize for PyDecoder {
+    fn serialize<S>(&self, _serializer: S) -> std::result::Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        Err(serde::ser::Error::custom(
+            "Custom PyDecoder cannot be serialized",
+        ))
+    }
+}
+
+impl<'de> Deserialize<'de> for PyDecoder {
+    fn deserialize<D>(_deserializer: D) -> std::result::Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        unimplemented!("PyDecoder cannot be deserialized")
     }
 }